refactor: :art: Removed some console.log + misc.
